feat: add decorator to Storybook preview
