+++ .storybook/vitest.setup.ts
@@ -1,6 +1,6 @@
 import * as a11yAddonAnnotations from "@storybook/addon-a11y/preview";
 import { setProjectAnnotations } from '@storybook/react-vite';
-import * as projectAnnotations from './preview';
+import * as projectAnnotations from './preview.tsx';
 
 // This is an important step to apply the right configuration when testing your stories.
 // More info at: https://storybook.js.org/docs/api/portable-stories/portable-stories-vitest#setprojectannotations
